* Make module out of recolor command
